+++ R/isoband.R
@@ -1,10 +1,4 @@
-#' Generate isolines and isobands
-#'
-#' @name isoband
-#' @docType package
 #' @useDynLib isoband, .registration = TRUE
-#' @importFrom Rcpp sourceCpp
 #' @importFrom utils modifyList
 #' @import grid
-NULL
-
+"_PACKAGE"