+++ src/test-runner.cpp
@@ -3,5 +3,6 @@
  * 'run_testthat_tests()' C routine that can be used to run the Catch unit tests
  * available in your package.
  */
+#define CATCH_CONFIG_RUNNER
 #define TESTTHAT_TEST_RUNNER
 #include <testthat.h>